Moved the logic of formDetails higher up in the component tree in create and edit contest components.
