+++ src/Components/superuser/createContest/ContestDetails.jsx
@@ -1,25 +1,18 @@
 /* eslint-disable no-param-reassign */
-import React, { useState } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { Cell, Grid, Row } from '@material/react-layout-grid';
 import Select, { Option } from '@material/react-select';
 import TextField, { Input } from '@material/react-text-field';
 import DatePicker from '../../common/DatePicker/index';
 import TimePicker from '../../common/TimePicker/index';
 import MultiSelect from '../../common/MultiSelect/index';
+import Editor from '../../common/MarkdownEditor/Editor';
+import EditorContainer from '../../common/MarkdownEditor/EditorContainer';
 
-const ContestDetails = () => {
-  const intialFormDetails = {
-    code: '',
-    type: 'round',
-    name: '',
-    description: '',
-    admins: [],
-    start: new Date(),
-    end: new Date(),
-    solutionVisibility: 'after',
-  };
+const ContestDetails = ({ formDetails, setFormDetails }) => {
   const options = ['ac030540@gmail.com', 'ninadc32@gmail.com', 'test@gmail.com', 'test2@gmail.com'];
-  const [formDetails, setFormDetails] = useState(intialFormDetails);
+
   // console.log(formDetails.admins);
   const onTextFieldChange = (event, keyToBeUpdated) => {
     const { value } = event.target;
@@ -36,9 +29,17 @@ const ContestDetails = () => {
     return { ...previousFormDetails };
   });
 
+
   const onSelectChange = (index, item, keyToBeUpdated) => (
     setFormDetails((previousFormDetails) => {
       previousFormDetails[keyToBeUpdated] = item.getAttribute('data-value');
+      return { ...previousFormDetails };
+    })
+  );
+
+  const updateEditorStates = (keyToBeUpdated, value) => (
+    setFormDetails((previousFormDetails) => {
+      previousFormDetails[keyToBeUpdated] = value;
       return { ...previousFormDetails };
     })
   );
@@ -99,16 +100,9 @@ const ContestDetails = () => {
       </Row>
       <Row>
         <Cell desktopColumns={12} tabletColumns={8} phoneColumns={4}>
-          <TextField
-            label="Description"
-            className="mb3 text-area-width-100"
-            textarea
-          >
-            <Input
-              value={formDetails.description}
-              onChange={e => onTextFieldChange(e, 'description')}
-            />
-          </TextField>
+          <EditorContainer title="Description">
+            <Editor value={formDetails.description} setValue={value => updateEditorStates('description', value)} />
+          </EditorContainer>
         </Cell>
       </Row>
       <Row>
@@ -155,6 +149,11 @@ const ContestDetails = () => {
       </Row>
     </Grid>
   );
+};
+
+ContestDetails.propTypes = {
+  formDetails: PropTypes.object.isRequired,
+  setFormDetails: PropTypes.func.isRequired,
 };
 
 export default ContestDetails;
